+++ modules/Material/ListItems/BaseListItem.qml
@@ -16,7 +16,8 @@ View {
     property int dividerInset: 0
     property bool showDivider: false
 
-    signal triggered()
+    signal clicked()
+    signal pressAndHold()
 
     ThinDivider {
         anchors.bottom: parent.bottom
@@ -27,11 +28,17 @@ View {
 
     Ink {
         id: ink
-        onClicked: listItem.triggered()
+
+        onClicked: listItem.clicked()
+        onPressAndHold: listItem.pressAndHold()
+
         anchors.fill: parent
+
         enabled: listItem.interactive
         z: -1
     }
 
-    tintColor: selected ? Qt.rgba(0,0,0,0.05) : ink.containsMouse ? Qt.rgba(0,0,0,0.03) : Qt.rgba(0,0,0,0)
+    tintColor: selected
+               ? Qt.rgba(0,0,0,0.05)
+               : ink.containsMouse ? Qt.rgba(0,0,0,0.03) : Qt.rgba(0,0,0,0)
 }