+++ modules/Material/ListItems/SimpleMenu.qml
@@ -9,7 +9,7 @@ Subtitled {
 
     subText: listView.currentItem.text
 
-    onTriggered: menu.open(listItem, units.dp(16), 0)
+    onClicked: menu.open(listItem, units.dp(16), 0)
 
     Dropdown {
         id: menu
@@ -37,7 +37,7 @@ Subtitled {
 
                 text: modelData
 
-                onTriggered: {
+                onClicked: {
                     listView.currentIndex = index
                     menu.close()
                 }
@@ -45,4 +45,3 @@ Subtitled {
         }
     }
 }
-