+++ CHANGELOG.md
@@ -3,6 +3,7 @@ QML Material CHANGELOG
 
  * Move popovers as necessary to prevent them being clipped by the window
  * [PageStack] Add support for a divided page and toolbar
+ * Rename list item and IconButton onTriggered to onClicked
  * Minor bug fixes and code improvements